veil-py/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
tesseract==0.1.3
Werkzeug==3.1.3
